Restructuring, completing tests, WIP
