+++ fundraiser/project/fundraiser-contract/tests/functions/core/cancel_campaign.rs
@@ -1,11 +1,12 @@
 use crate::utils::{
-    abi_calls::{campaign_info, cancel_campaign, create_campaign},
-    test_helpers::setup,
+    interface::core::{cancel_campaign, create_campaign},
+    setup::{setup, CancelledCampaignEvent},
 };
 
 mod success {
 
     use super::*;
+    use crate::utils::{interface::info::campaign_info, setup::State};
 
     #[tokio::test]
     async fn cancels() {
@@ -20,17 +21,22 @@ mod success {
         )
         .await;
 
-        assert_eq!(
-            false,
-            campaign_info(&author.contract, 1).await.value.cancelled
-        );
+        assert!(matches!(
+            campaign_info(&author.contract, 1).await.value.state,
+            State::Funding()
+        ));
 
-        cancel_campaign(&author.contract, 1).await;
+        let response = cancel_campaign(&author.contract, 1).await;
 
-        assert_eq!(
-            true,
-            campaign_info(&author.contract, 1).await.value.cancelled
-        );
+        let log = response.get_logs_with_type::<CancelledCampaignEvent>().unwrap();
+        let event = log.get(0).unwrap();
+
+        assert!(matches!(
+            campaign_info(&author.contract, 1).await.value.state,
+            State::Cancelled()
+        ));
+
+        assert_eq!(*event, CancelledCampaignEvent { id: 1 });
     }
 
     #[tokio::test]
@@ -45,7 +51,6 @@ mod success {
             defaults.target_amount,
         )
         .await;
-
         create_campaign(
             &author.contract,
             &defaults.asset_id,
@@ -55,29 +60,26 @@ mod success {
         )
         .await;
 
-        assert_eq!(
-            false,
-            campaign_info(&author.contract, 1).await.value.cancelled
-        );
-
-        assert_eq!(
-            false,
-            campaign_info(&author.contract, 2).await.value.cancelled
-        );
+        assert!(matches!(
+            campaign_info(&author.contract, 1).await.value.state,
+            State::Funding()
+        ));
+        assert!(matches!(
+            campaign_info(&author.contract, 2).await.value.state,
+            State::Funding()
+        ));
 
         cancel_campaign(&author.contract, 1).await;
-
-        assert_eq!(
-            true,
-            campaign_info(&author.contract, 1).await.value.cancelled
-        );
+        assert!(matches!(
+            campaign_info(&author.contract, 1).await.value.state,
+            State::Cancelled()
+        ));
 
         cancel_campaign(&author.contract, 2).await;
-
-        assert_eq!(
-            true,
-            campaign_info(&author.contract, 2).await.value.cancelled
-        );
+        assert!(matches!(
+            campaign_info(&author.contract, 2).await.value.state,
+            State::Cancelled()
+        ));
     }
 }
 
@@ -86,7 +88,7 @@ mod revert {
     use super::*;
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "InvalidID")]
     async fn when_id_is_zero() {
         let (author, _, _, _, defaults) = setup().await;
 
@@ -104,7 +106,7 @@ mod revert {
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "InvalidID")]
     async fn when_id_is_greater_than_number_of_campaigns() {
         let (author, _, _, _, _) = setup().await;
 
@@ -113,10 +115,11 @@ mod revert {
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "UnauthorizedUser")]
     async fn when_sender_is_not_author() {
         let (author, user, _, _, defaults) = setup().await;
-        let deadline = 4;
+        let provider = author.wallet.get_provider().unwrap();
+        let deadline = provider.latest_block_height().await.unwrap() + 2;
 
         create_campaign(
             &author.contract,
@@ -132,10 +135,11 @@ mod revert {
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "CampaignEnded")]
     async fn when_calling_after_deadline() {
         let (author, _, _, _, defaults) = setup().await;
-        let deadline = 3;
+        let provider = author.wallet.get_provider().unwrap();
+        let deadline = provider.latest_block_height().await.unwrap() + 2;
 
         create_campaign(
             &author.contract,
@@ -151,7 +155,7 @@ mod revert {
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "CampaignHasBeenCancelled")]
     async fn when_calling_after_already_cancelled() {
         let (author, _, _, _, defaults) = setup().await;
