+++ fundraiser/project/fundraiser-contract/tests/functions/core/mod.rs
@@ -1,0 +1,5 @@
+mod cancel_campaign;
+mod claim_pledges;
+mod create_campaign;
+mod pledge;
+mod unpledge;
