+++ fundraiser/project/fundraiser-contract/tests/functions/core/pledge.rs
@@ -1,15 +1,16 @@
 use crate::utils::{
-    abi_calls::{
-        asset_info_by_count, campaign_info, cancel_campaign, create_campaign, pledge, pledge_count,
-        pledged,
-    },
-    test_helpers::{identity, mint, setup},
+    interface::core::{cancel_campaign, create_campaign, pledge},
+    setup::{mint, setup},
 };
-use fuels::tx::AssetId;
 
 mod success {
 
     use super::*;
+    use crate::utils::{
+        interface::info::{asset_info_by_count, campaign_info, pledge_count, pledged},
+        setup::identity,
+    };
+    use fuels::tx::AssetId;
 
     #[tokio::test]
     async fn pledges() {
@@ -51,7 +52,6 @@ mod success {
         );
 
         pledge(&user.contract, 1, &asset, defaults.target_amount).await;
-
         assert_eq!(
             0,
             user.wallet
@@ -122,7 +122,6 @@ mod success {
         );
 
         pledge(&user.contract, 1, &asset, defaults.target_amount).await;
-
         assert_eq!(
             defaults.target_amount,
             user.wallet
@@ -152,7 +151,6 @@ mod success {
         assert_eq!(defaults.target_amount, info.amount);
 
         pledge(&user.contract, 1, &asset, defaults.target_amount).await;
-
         assert_eq!(
             0,
             user.wallet
@@ -216,7 +214,6 @@ mod success {
             defaults.target_amount,
         )
         .await;
-
         create_campaign(
             &author.contract,
             &asset2_id,
@@ -246,7 +243,6 @@ mod success {
                 .await
                 .unwrap()
         );
-
         assert_eq!(
             defaults.target_amount,
             user.wallet
@@ -329,7 +325,7 @@ mod revert {
     use super::*;
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "InvalidID")]
     async fn when_id_is_zero() {
         let (author, user, asset, _, defaults) = setup().await;
 
@@ -353,7 +349,7 @@ mod revert {
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "InvalidID")]
     async fn when_id_is_greater_than_number_of_campaigns() {
         let (author, user, asset, _, defaults) = setup().await;
 
@@ -377,11 +373,11 @@ mod revert {
     }
 
     #[tokio::test]
-    #[ignore]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "CampaignEnded")]
     async fn when_pledging_after_deadline() {
         let (author, user, asset, _, defaults) = setup().await;
-        let deadline = 5;
+        let provider = author.wallet.get_provider().unwrap();
+        let deadline = provider.latest_block_height().await.unwrap() + 3;
 
         mint(
             &asset.contract,
@@ -399,14 +395,12 @@ mod revert {
         .await;
         pledge(&user.contract, 1, &asset, defaults.target_amount).await;
 
-        // TODO: shift block height to be after deadline
-
         // Reverts
         pledge(&user.contract, 1, &asset, 0).await;
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "IncorrectAssetSent")]
     async fn when_pledging_incorrect_asset() {
         let (author, user, _, asset2, defaults) = setup().await;
 
@@ -416,7 +410,6 @@ mod revert {
             user.wallet.address(),
         )
         .await;
-
         create_campaign(
             &author.contract,
             &defaults.asset_id,
@@ -431,7 +424,7 @@ mod revert {
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "AmountCannotBeZero")]
     async fn when_pledging_zero_amount() {
         let (author, user, asset, _, defaults) = setup().await;
 
@@ -441,7 +434,6 @@ mod revert {
             user.wallet.address(),
         )
         .await;
-
         create_campaign(
             &author.contract,
             &defaults.asset_id,
@@ -456,10 +448,11 @@ mod revert {
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "CampaignHasBeenCancelled")]
     async fn when_pledging_to_cancelled_campaign() {
         let (author, user, asset, _, defaults) = setup().await;
-        let deadline = 5;
+        let provider = author.wallet.get_provider().unwrap();
+        let deadline = provider.latest_block_height().await.unwrap() + 5;
 
         mint(
             &asset.contract,
@@ -475,7 +468,6 @@ mod revert {
             defaults.target_amount,
         )
         .await;
-
         cancel_campaign(&author.contract, 1).await;
 
         // Reverts
