+++ fundraiser/project/fundraiser-contract/tests/functions/info/campaign_info.rs
@@ -1,11 +1,12 @@
 use crate::utils::{
-    abi_calls::{campaign_info, create_campaign},
-    test_helpers::{identity, setup},
+    interface::{core::create_campaign, info::campaign_info},
+    setup::setup,
 };
 
 mod success {
 
     use super::*;
+    use crate::utils::setup::{identity, State};
 
     #[tokio::test]
     async fn returns_info() {
@@ -25,8 +26,7 @@ mod success {
         assert_eq!(info.asset, defaults.asset_id);
         assert_eq!(info.author, identity(author.wallet.address()).await);
         assert_eq!(info.beneficiary, defaults.beneficiary);
-        assert_eq!(info.cancelled, false);
-        assert_eq!(info.claimed, false);
+        assert_eq!(info.state, State::Funding());
         assert_eq!(info.deadline, defaults.deadline);
         assert_eq!(info.target_amount, defaults.target_amount);
         assert_eq!(info.total_pledge, 0);
@@ -38,7 +38,7 @@ mod revert {
     use super::*;
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "InvalidID")]
     async fn when_id_is_zero() {
         let (author, _, _, _, _) = setup().await;
 
@@ -47,7 +47,7 @@ mod revert {
     }
 
     #[tokio::test]
-    #[should_panic(expected = "Revert(18446744073709486080)")]
+    #[should_panic(expected = "InvalidID")]
     async fn when_id_is_greater_than_number_of_campaigns() {
         let (author, _, _, _, _) = setup().await;
