+++ fundraiser/project/fundraiser-contract/tests/functions/info/mod.rs
@@ -1,0 +1,9 @@
+mod asset_count;
+mod asset_info_by_count;
+mod asset_info_by_id;
+mod campaign;
+mod campaign_info;
+mod pledge_count;
+mod pledged;
+mod total_campaigns;
+mod user_campaign_count;
