+++ fundraiser/project/fundraiser-contract/tests/functions/info/total_campaigns.rs
@@ -1,24 +1,22 @@
-use crate::utils::{
-    abi_calls::{asset_count, create_campaign},
-    test_helpers::setup,
-};
-
 mod success {
 
-    use super::*;
+    use crate::utils::{
+        interface::{core::create_campaign, info::total_campaigns},
+        setup::setup,
+    };
 
     #[tokio::test]
     async fn returns_zero() {
         let (author, _, _, _, _) = setup().await;
 
-        assert_eq!(0, asset_count(&author.contract).await);
+        assert_eq!(0, total_campaigns(&author.contract).await);
     }
 
     #[tokio::test]
     async fn returns_one() {
         let (author, _, _, _, defaults) = setup().await;
 
-        assert_eq!(0, asset_count(&author.contract).await);
+        assert_eq!(0, total_campaigns(&author.contract).await);
         create_campaign(
             &author.contract,
             &defaults.asset_id,
@@ -27,6 +25,6 @@ mod success {
             defaults.target_amount,
         )
         .await;
-        assert_eq!(1, asset_count(&author.contract).await);
+        assert_eq!(1, total_campaigns(&author.contract).await);
     }
 }
